fix: securityUserCacheManager Bean 정의
