+++ src/main/java/com/kcy/fitapet/global/config/redis/SecurityUserCacheManager.java
@@ -1,0 +1,13 @@
+package com.kcy.fitapet.global.config.redis;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD,
+        ElementType.TYPE, ElementType.ANNOTATION_TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Qualifier("securityUserCacheManager")
+public @interface SecurityUserCacheManager {
+}
